backend/requirements.txt
chromadb==1.1.1
fastapi==0.115.12
google-generativeai==0.8.5
python-multipart==0.0.20
sentence-transformers==3.4.1
trimesh==4.6.8
uvicorn[standard]==0.34.0
